Fix display of map properties to be shown in sceneCoordinates

+++ ogm_gui/include/ogm_gui/ogm_gui_controller.h
@@ -183,9 +183,9 @@ namespace ogm_gui
       **/
       void updateMapInternal(void);
 
-      void moveMapHorizontally(int x);
+      void moveMapHorizontally(double x);
 
-      void moveMapVertically(int y);
+      void moveMapVertically(double y);
 
       void rotateMap(int r);
